+++ packages/telescope-invites/lib/server/publications.js
@@ -1,0 +1,4 @@
+Meteor.publish('invites', function (userId) {
+  var invites = Invites.find({invitingUserId: userId})
+  return (this.userId === userId || isAdmin(Meteor.user())) ? invites : []
+});